requirements.txt
# VEO3 Consistency Generator Requirements

# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0

# API clients
google-generativeai==0.3.2

# Data handling
pyyaml==6.0.1

# Web framework (optional, for future API)
fastapi==0.104.1
uvicorn==0.24.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
python-dateutil==2.8.2
